requirements.txt
fpdf==1.7.2
numpy==1.21.2
Pillow==8.3.2
smbus==1.1.post2
spidev==3.5
rpi-ws281x=5.0.0
wxPython==4.2.1